chore(backend): delete all useless tables
