--- apps/backend/src/entities/book.rs
@@ -1,35 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.11.2
-
-use sea_orm::entity::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use crate::migrator::BookSource;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize, Default)]
-#[sea_orm(table_name = "book")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub metadata_id: i32,
-    pub num_pages: Option<i32>,
-    pub source: BookSource,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::metadata::Entity",
-        from = "Column::MetadataId",
-        to = "super::metadata::Column::Id",
-        on_update = "Cascade",
-        on_delete = "Cascade"
-    )]
-    Metadata,
-}
-
-impl Related<super::metadata::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Metadata.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
