--- apps/backend/src/entities/movie.rs
@@ -1,35 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.11.2
-
-use sea_orm::entity::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use crate::migrator::MovieSource;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize, Default)]
-#[sea_orm(table_name = "movie")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub metadata_id: i32,
-    pub runtime: Option<i32>,
-    pub source: MovieSource,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::metadata::Entity",
-        from = "Column::MetadataId",
-        to = "super::metadata::Column::Id",
-        on_update = "Cascade",
-        on_delete = "Cascade"
-    )]
-    Metadata,
-}
-
-impl Related<super::metadata::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Metadata.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
