+++ apps/backend/src/entities/prelude.rs
@@ -1,19 +1,13 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.11.3
 
-pub use super::audio_book::Entity as AudioBook;
-pub use super::book::Entity as Book;
 pub use super::collection::Entity as Collection;
 pub use super::genre::Entity as Genre;
 pub use super::media_import_report::Entity as MediaImportReport;
 pub use super::metadata::Entity as Metadata;
 pub use super::metadata_to_collection::Entity as MetadataToCollection;
 pub use super::metadata_to_genre::Entity as MetadataToGenre;
-pub use super::movie::Entity as Movie;
-pub use super::podcast::Entity as Podcast;
 pub use super::review::Entity as Review;
 pub use super::seen::Entity as Seen;
-pub use super::show::Entity as Show;
 pub use super::summary::Entity as Summary;
 pub use super::user::Entity as User;
 pub use super::user_to_metadata::Entity as UserToMetadata;
-pub use super::video_game::Entity as VideoGame;
