--- apps/backend/src/entities/show.rs
@@ -1,35 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.11.2
-
-use sea_orm::entity::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use crate::{migrator::ShowSource, shows::ShowSpecifics};
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize, Default)]
-#[sea_orm(table_name = "show")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub metadata_id: i32,
-    pub details: ShowSpecifics,
-    pub source: ShowSource,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::metadata::Entity",
-        from = "Column::MetadataId",
-        to = "super::metadata::Column::Id",
-        on_update = "Cascade",
-        on_delete = "Cascade"
-    )]
-    Metadata,
-}
-
-impl Related<super::metadata::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Metadata.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
