--- apps/backend/src/entities/token.rs
@@ -1,39 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.11.2
-
-use sea_orm::entity::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use crate::migrator::TokenLot;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
-#[sea_orm(table_name = "token")]
-pub struct Model {
-    #[sea_orm(primary_key)]
-    pub id: i32,
-    #[sea_orm(indexed, unique)]
-    pub value: String,
-    pub lot: TokenLot,
-    pub user_id: i32,
-    pub created_on: DateTimeUtc,
-    pub last_used: Option<DateTimeUtc>,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::UserId",
-        to = "super::user::Column::Id",
-        on_update = "Cascade",
-        on_delete = "Cascade"
-    )]
-    User,
-}
-
-impl Related<super::user::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::User.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
