+++ apps/backend/src/migrator/m20230410_000001_create_metadata.rs
@@ -17,6 +17,39 @@ pub enum MetadataImageLot {
     Poster,
 }
 
+// The different types of media that can be stored
+#[derive(
+    Debug,
+    Clone,
+    Copy,
+    PartialEq,
+    Eq,
+    EnumIter,
+    DeriveActiveEnum,
+    Deserialize,
+    Serialize,
+    Enum,
+    Default,
+)]
+#[sea_orm(rs_type = "String", db_type = "String(None)")]
+pub enum MetadataSource {
+    #[default]
+    #[sea_orm(string_value = "AU")]
+    Audible,
+    #[sea_orm(string_value = "CU")]
+    Custom,
+    #[sea_orm(string_value = "GO")]
+    Goodreads,
+    #[sea_orm(string_value = "IG")]
+    Igdb,
+    #[sea_orm(string_value = "LI")]
+    Listennotes,
+    #[sea_orm(string_value = "OL")]
+    Openlibrary,
+    #[sea_orm(string_value = "TM")]
+    Tmdb,
+}
+
 // This is responsible for storing common metadata about all media items
 #[derive(Iden)]
 pub enum MetadataImage {
