--- apps/backend/src/migrator/m20230502_000008_create_video_game.rs
@@ -1,78 +1,0 @@
-use async_graphql::Enum;
-use sea_orm::{DeriveActiveEnum, EnumIter};
-use sea_orm_migration::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use super::Metadata;
-
-pub struct Migration;
-
-#[derive(
-    Debug,
-    Clone,
-    Copy,
-    PartialEq,
-    Eq,
-    EnumIter,
-    DeriveActiveEnum,
-    Deserialize,
-    Serialize,
-    Enum,
-    Default,
-)]
-#[sea_orm(rs_type = "String", db_type = "String(None)")]
-pub enum VideoGameSource {
-    #[default]
-    #[sea_orm(string_value = "I")]
-    Igdb,
-    #[sea_orm(string_value = "C")]
-    Custom,
-}
-
-#[derive(Iden)]
-pub enum VideoGame {
-    Table,
-    MetadataId,
-    Source,
-    Details,
-}
-
-impl MigrationName for Migration {
-    fn name(&self) -> &str {
-        "m20230502_000008_create_video_game"
-    }
-}
-
-#[async_trait::async_trait]
-impl MigrationTrait for Migration {
-    async fn up(&self, manager: &SchemaManager) -> Result<(), DbErr> {
-        manager
-            .create_table(
-                Table::create()
-                    .table(VideoGame::Table)
-                    .col(
-                        ColumnDef::new(VideoGame::MetadataId)
-                            .integer()
-                            .primary_key()
-                            .unique_key()
-                            .not_null(),
-                    )
-                    .foreign_key(
-                        ForeignKey::create()
-                            .name("video_game_to_metadata_foreign_key")
-                            .from(VideoGame::Table, VideoGame::MetadataId)
-                            .to(Metadata::Table, Metadata::Id)
-                            .on_delete(ForeignKeyAction::Cascade)
-                            .on_update(ForeignKeyAction::Cascade),
-                    )
-                    .col(ColumnDef::new(VideoGame::Source).string_len(1).not_null())
-                    .to_owned(),
-            )
-            .await?;
-        Ok(())
-    }
-
-    async fn down(&self, _manager: &SchemaManager) -> Result<(), DbErr> {
-        Ok(())
-    }
-}
